Repair several bugs
B	There is no need to read octets in BGPCommonHeader if the remaining length is zero
B	Correction of Extended Communities attribute format (was incorrectly parsed)
B	Correction of printing IPv6 address in MPReachNLRI/MPUnreachNLRI

+++ include/AttributeTypeExtCommunities.h
@@ -40,8 +40,8 @@ using namespace std;
 
 typedef struct ExtCommunityValue
 {
-	uint16_t	typeHigh;
-	uint16_t	typeLow;
+	uint8_t		typeHigh;
+	uint8_t		typeLow;
 	uint8_t		rchValue[6];
 } ExtCommunityValue;
